+++ core/src/main/kotlin/no/elg/hex/screens/PlayableIslandScreen.kt
@@ -46,6 +46,7 @@ import no.elg.hex.input.GameInputProcessor
 import no.elg.hex.island.Island
 import no.elg.hex.island.Territory
 import no.elg.hex.renderer.DebugGraphRenderer
+import no.elg.hex.renderer.StrengthBarRenderer
 import no.elg.hex.screens.LevelSelectScreen.PreviewModifier
 import no.elg.hex.screens.LevelSelectScreen.PreviewModifier.LOST
 import no.elg.hex.screens.LevelSelectScreen.PreviewModifier.NOTHING
@@ -55,6 +56,7 @@ import no.elg.hex.util.canAttack
 import no.elg.hex.util.createHandInstance
 import no.elg.hex.util.getData
 import no.elg.hex.util.hide
+import no.elg.hex.util.isLazyInitialized
 import no.elg.hex.util.onAnyKeysDownEvent
 import no.elg.hex.util.onInteract
 import no.elg.hex.util.show
@@ -69,6 +71,7 @@ class PlayableIslandScreen(id: Int, island: Island) : PreviewIslandScreen(id, is
 
   private val frameUpdatable by lazy { GameInfoRenderer(this) }
   private val debugRenderer: DebugInfoRenderer by lazy { DebugInfoRenderer(this) }
+  private val strengthBarRenderer by lazy { StrengthBarRenderer(this.island) }
 
   private val confirmEndTurn: VisWindow
   private val confirmSurrender: VisWindow
@@ -483,6 +486,9 @@ class PlayableIslandScreen(id: Int, island: Island) : PreviewIslandScreen(id, is
     if (DebugGraphRenderer.isEnabled) {
       DebugGraphRenderer.frameUpdate()
     }
+    if (StrengthBarRenderer.isEnabled) {
+      strengthBarRenderer.frameUpdate()
+    }
     debugRenderer.frameUpdate()
     frameUpdatable.frameUpdate()
     stageScreen.render(delta)
@@ -498,6 +504,9 @@ class PlayableIslandScreen(id: Int, island: Island) : PreviewIslandScreen(id, is
     frameUpdatable.resize(width, height)
     if (DebugGraphRenderer.isEnabled) {
       DebugGraphRenderer.resize(width, height)
+    }
+    if (StrengthBarRenderer.isEnabled) {
+      strengthBarRenderer.resize(width, height)
     }
   }
 
@@ -529,6 +538,10 @@ class PlayableIslandScreen(id: Int, island: Island) : PreviewIslandScreen(id, is
     listener.disposeSafely()
 
     DebugGraphRenderer.dispose()
+
+    if (::strengthBarRenderer.isLazyInitialized) {
+      strengthBarRenderer.dispose()
+    }
   }
 
   companion object {